requirements.txt
Flask==2.3.2
line-bot-sdk==3.9.0
gunicorn==20.0.4